+++ extensions/binding/org.eclipse.smarthome.binding.digitalstrom/src/main/java/org/eclipse/smarthome/binding/digitalstrom/internal/lib/structure/scene/SceneDiscovery.java
@@ -267,7 +267,7 @@ public class SceneDiscovery {
                                                 groupIdInter = null;
                                             }
                                             if (groupID != null) {
-                                                if (FunctionalColorGroupEnum.getColorGroup((int) groupID)
+                                                if (FunctionalColorGroupEnum.getColorGroup(groupID)
                                                         .equals(FunctionalColorGroupEnum.YELLOW)) {
                                                     discoverScene(SceneEnum.AUTO_OFF.getSceneNumber(), groupID);
                                                 }
